Convert board & card components
